windows_registry

Struct Key

pub struct Key(/* private fields */);

A registry key.

Implementations

impl Key

pub fn create<T: AsRef<str>>(&self, path: T) -> Result<Self>

Creates a registry key. If the key already exists, the function opens it.

pub fn open<T: AsRef<str>>(&self, path: T) -> Result<Self>

Opens a registry key.

pub fn options(&self) -> OpenOptions<'_>

Creates an OpenOptions object for the registry key.

pub unsafe fn from_raw(handle: *mut c_void) -> Self

Constructs a registry key from an existing handle.

Safety

This function takes ownership of the handle. The handle must be owned by the caller and safe to free with RegCloseKey.

pub fn as_raw(&self) -> *mut c_void

Returns the underlying registry key handle.

pub fn rename<F: AsRef<str>, T: AsRef<str>>(&self, from: F, to: T) -> Result<()>

Changes the name of the specified registry key.

pub fn remove_tree<T: AsRef<str>>(&self, path: T) -> Result<()>

Removes the registry keys and values of the specified key recursively.

pub fn remove_value<T: AsRef<str>>(&self, name: T) -> Result<()>

Removes the registry value.

pub fn keys(&self) -> Result<KeyIterator<'_>>

Creates an iterator of registry key names.

pub fn values(&self) -> Result<ValueIterator<'_>>

Creates an iterator of registry values.

pub fn set_u32<T: AsRef<str>>(&self, name: T, value: u32) -> Result<()>

Sets the name and value in the registry key.

pub fn set_u64<T: AsRef<str>>(&self, name: T, value: u64) -> Result<()>

Sets the name and value in the registry key.

pub fn set_string<T: AsRef<str>, U: AsRef<str>>( &self, name: T, value: U, ) -> Result<()>

Sets the name and value in the registry key.

pub fn set_hstring<T: AsRef<str>>(&self, name: T, value: &HSTRING) -> Result<()>

Sets the name and value in the registry key.

pub fn set_expand_string<T: AsRef<str>, U: AsRef<str>>( &self, name: T, value: U, ) -> Result<()>

Sets the name and value in the registry key.

pub fn set_expand_hstring<T: AsRef<str>>( &self, name: T, value: &HSTRING, ) -> Result<()>

Sets the name and value in the registry key.

pub fn set_multi_string<T: AsRef<str>>( &self, name: T, value: &[T], ) -> Result<()>

Sets the name and value in the registry key.

pub fn set_value<T: AsRef<str>>(&self, name: T, value: &Value) -> Result<()>

Sets the name and value in the registry key.

pub fn set_bytes<T: AsRef<str>>( &self, name: T, ty: Type, value: &[u8], ) -> Result<()>

Sets the name and value in the registry key.

pub fn get_type<T: AsRef<str>>(&self, name: T) -> Result<Type>

Gets the type for the name in the registry key.

pub fn get_value<T: AsRef<str>>(&self, name: T) -> Result<Value>

Gets the value for the name in the registry key.

pub fn get_u32<T: AsRef<str>>(&self, name: T) -> Result<u32>

Gets the value for the name in the registry key.

pub fn get_u64<T: AsRef<str>>(&self, name: T) -> Result<u64>

Gets the value for the name in the registry key.

pub fn get_string<T: AsRef<str>>(&self, name: T) -> Result<String>

Gets the value for the name in the registry key.

pub fn get_hstring<T: AsRef<str>>(&self, name: T) -> Result<HSTRING>

Gets the value for the name in the registry key.

pub fn get_multi_string<T: AsRef<str>>(&self, name: T) -> Result<Vec<String>>

Gets the value for the name in the registry key.

pub unsafe fn raw_set_bytes<N: AsRef<PCWSTR>>( &self, name: N, ty: Type, value: &[u8], ) -> Result<()>

Sets the name and value in the registry key.

This method avoids any allocations.

Safety

The PCWSTR pointer needs to be valid for reads up until and including the next \0.

pub unsafe fn raw_get_info<N: AsRef<PCWSTR>>( &self, name: N, ) -> Result<(Type, usize)>

Gets the type and length for the name in the registry key.

This method avoids any allocations.

Safety

The PCWSTR pointer needs to be valid for reads up until and including the next \0.

pub unsafe fn raw_get_bytes<'a, N: AsRef<PCWSTR>>( &self, name: N, value: &'a mut [u8], ) -> Result<(Type, &'a [u8])>

Gets the value for the name in the registry key.

This method avoids any allocations.

Safety

The PCWSTR pointer needs to be valid for reads up until and including the next \0.